+++ lib/processor/compress.go
@@ -39,15 +39,9 @@ func init() {
 		constructor: NewCompress,
 		description: `
 Compresses parts of a message according to the selected algorithm. Supported
-available compression types are: gzip, zlib, flate. If the list of target parts
-is empty the compression will be applied to all message parts.
+available compression types are: gzip, zlib, flate.
 
-The 'level' field might not apply to all algorithms.
-
-Part indexes can be negative, and if so the part will be selected from the end
-counting backwards starting from -1. E.g. if index = -1 then the selected part
-will be the last part of the message, if index = -2 then the part before the
-last element with be selected, and so on.`,
+The 'level' field might not apply to all algorithms.`,
 	}
 }