+++ lib/processor/constructor.go
@@ -55,59 +55,61 @@ var Constructors = map[string]TypeSpec{}
 
 // Config is the all encompassing configuration struct for all processor types.
 type Config struct {
-	Type        string            `json:"type" yaml:"type"`
-	Archive     ArchiveConfig     `json:"archive" yaml:"archive"`
-	Batch       BatchConfig       `json:"batch" yaml:"batch"`
-	BoundsCheck BoundsCheckConfig `json:"bounds_check" yaml:"bounds_check"`
-	Combine     CombineConfig     `json:"combine" yaml:"combine"`
-	Compress    CompressConfig    `json:"compress" yaml:"compress"`
-	Conditional ConditionalConfig `json:"conditional" yaml:"conditional"`
-	Decode      DecodeConfig      `json:"decode" yaml:"decode"`
-	Decompress  DecompressConfig  `json:"decompress" yaml:"decompress"`
-	Dedupe      DedupeConfig      `json:"dedupe" yaml:"dedupe"`
-	Encode      EncodeConfig      `json:"encode" yaml:"encode"`
-	Filter      FilterConfig      `json:"filter" yaml:"filter"`
-	FilterParts FilterPartsConfig `json:"filter_parts" yaml:"filter_parts"`
-	Grok        GrokConfig        `json:"grok" yaml:"grok"`
-	HashSample  HashSampleConfig  `json:"hash_sample" yaml:"hash_sample"`
-	HTTP        HTTPConfig        `json:"http" yaml:"http"`
-	InsertPart  InsertPartConfig  `json:"insert_part" yaml:"insert_part"`
-	JMESPath    JMESPathConfig    `json:"jmespath" yaml:"jmespath"`
-	JSON        JSONConfig        `json:"json" yaml:"json"`
-	MergeJSON   MergeJSONConfig   `json:"merge_json" yaml:"merge_json"`
-	Sample      SampleConfig      `json:"sample" yaml:"sample"`
-	SelectParts SelectPartsConfig `json:"select_parts" yaml:"select_parts"`
-	Split       struct{}          `json:"split" yaml:"split"`
-	Unarchive   UnarchiveConfig   `json:"unarchive" yaml:"unarchive"`
+	Type         string             `json:"type" yaml:"type"`
+	Archive      ArchiveConfig      `json:"archive" yaml:"archive"`
+	Batch        BatchConfig        `json:"batch" yaml:"batch"`
+	BoundsCheck  BoundsCheckConfig  `json:"bounds_check" yaml:"bounds_check"`
+	Combine      CombineConfig      `json:"combine" yaml:"combine"`
+	Compress     CompressConfig     `json:"compress" yaml:"compress"`
+	Conditional  ConditionalConfig  `json:"conditional" yaml:"conditional"`
+	Decode       DecodeConfig       `json:"decode" yaml:"decode"`
+	Decompress   DecompressConfig   `json:"decompress" yaml:"decompress"`
+	Dedupe       DedupeConfig       `json:"dedupe" yaml:"dedupe"`
+	Encode       EncodeConfig       `json:"encode" yaml:"encode"`
+	Filter       FilterConfig       `json:"filter" yaml:"filter"`
+	FilterParts  FilterPartsConfig  `json:"filter_parts" yaml:"filter_parts"`
+	Grok         GrokConfig         `json:"grok" yaml:"grok"`
+	HashSample   HashSampleConfig   `json:"hash_sample" yaml:"hash_sample"`
+	HTTP         HTTPConfig         `json:"http" yaml:"http"`
+	InsertPart   InsertPartConfig   `json:"insert_part" yaml:"insert_part"`
+	JMESPath     JMESPathConfig     `json:"jmespath" yaml:"jmespath"`
+	JSON         JSONConfig         `json:"json" yaml:"json"`
+	MergeJSON    MergeJSONConfig    `json:"merge_json" yaml:"merge_json"`
+	ProcessField ProcessFieldConfig `json:"process_field" yaml:"process_field"`
+	Sample       SampleConfig       `json:"sample" yaml:"sample"`
+	SelectParts  SelectPartsConfig  `json:"select_parts" yaml:"select_parts"`
+	Split        struct{}           `json:"split" yaml:"split"`
+	Unarchive    UnarchiveConfig    `json:"unarchive" yaml:"unarchive"`
 }
 
 // NewConfig returns a configuration struct fully populated with default values.
 func NewConfig() Config {
 	return Config{
-		Type:        "bounds_check",
-		Archive:     NewArchiveConfig(),
-		Batch:       NewBatchConfig(),
-		BoundsCheck: NewBoundsCheckConfig(),
-		Combine:     NewCombineConfig(),
-		Compress:    NewCompressConfig(),
-		Conditional: NewConditionalConfig(),
-		Decode:      NewDecodeConfig(),
-		Decompress:  NewDecompressConfig(),
-		Dedupe:      NewDedupeConfig(),
-		Encode:      NewEncodeConfig(),
-		Filter:      NewFilterConfig(),
-		FilterParts: NewFilterPartsConfig(),
-		Grok:        NewGrokConfig(),
-		HashSample:  NewHashSampleConfig(),
-		HTTP:        NewHTTPConfig(),
-		InsertPart:  NewInsertPartConfig(),
-		JMESPath:    NewJMESPathConfig(),
-		JSON:        NewJSONConfig(),
-		MergeJSON:   NewMergeJSONConfig(),
-		Sample:      NewSampleConfig(),
-		SelectParts: NewSelectPartsConfig(),
-		Split:       struct{}{},
-		Unarchive:   NewUnarchiveConfig(),
+		Type:         "bounds_check",
+		Archive:      NewArchiveConfig(),
+		Batch:        NewBatchConfig(),
+		BoundsCheck:  NewBoundsCheckConfig(),
+		Combine:      NewCombineConfig(),
+		Compress:     NewCompressConfig(),
+		Conditional:  NewConditionalConfig(),
+		Decode:       NewDecodeConfig(),
+		Decompress:   NewDecompressConfig(),
+		Dedupe:       NewDedupeConfig(),
+		Encode:       NewEncodeConfig(),
+		Filter:       NewFilterConfig(),
+		FilterParts:  NewFilterPartsConfig(),
+		Grok:         NewGrokConfig(),
+		HashSample:   NewHashSampleConfig(),
+		HTTP:         NewHTTPConfig(),
+		InsertPart:   NewInsertPartConfig(),
+		JMESPath:     NewJMESPathConfig(),
+		JSON:         NewJSONConfig(),
+		MergeJSON:    NewMergeJSONConfig(),
+		ProcessField: NewProcessFieldConfig(),
+		Sample:       NewSampleConfig(),
+		SelectParts:  NewSelectPartsConfig(),
+		Split:        struct{}{},
+		Unarchive:    NewUnarchiveConfig(),
 	}
 }
 
@@ -171,9 +173,9 @@ func (m *Config) UnmarshalYAML(unmarshal func(interface{}) error) error {
 
 var header = "This document was generated with `benthos --list-processors`." + `
 
-Benthos has a concept of processors, these are functions that will be applied to
-each message passing through a pipeline. The function signature allows a
-processor to mutate or drop messages depending on the content of the message.
+Benthos processors are functions that will be applied to each message passing
+through a pipeline. The function signature allows a processor to mutate or drop
+messages depending on the content of the message.
 
 Processors are set via config, and depending on where in the config they are
 placed they will be run either immediately after a specific input (set in the
@@ -181,7 +183,23 @@ input section), on all messages (set in the pipeline section) or before a
 specific output (set in the output section).
 
 By organising processors you can configure complex behaviours in your pipeline.
-You can [find some examples here][0].`
+You can [find some examples here][0].
+
+### Batching and Multiple Part Messages
+
+All Benthos processors support mulitple part messages, which are synonymous with
+batches. Some processors such as [combine](#combine), [batch](#batch) and
+[split](#split) are able to create, expand and break down batches.
+
+Many processors are able to perform their behaviours on specific parts of a
+message batch, or on all parts, and have a field ` + "`parts`" + ` for
+specifying and array of part indexes they should apply to. If the list of target
+parts is empty these processors will be applied to all message parts.
+
+Part indexes can be negative, and if so the part will be selected from the end
+counting backwards starting from -1. E.g. if part = -1 then the selected part
+will be the last part of the message, if part = -2 then the part before the last
+element with be selected, and so on.`
 
 var footer = `
 [0]: ./examples.md`