+++ lib/processor/decompress.go
@@ -41,13 +41,7 @@ func init() {
 		constructor: NewDecompress,
 		description: `
 Decompresses the parts of a message according to the selected algorithm.
-Supported decompression types are: gzip, zlib, bzip2, flate. If the list of
-target parts is empty the decompression will be applied to all message parts.
-
-Part indexes can be negative, and if so the part will be selected from the end
-counting backwards starting from -1. E.g. if index = -1 then the selected part
-will be the last part of the message, if index = -2 then the part before the
-last element with be selected, and so on.
+Supported decompression types are: gzip, zlib, bzip2, flate.
 
 Parts that fail to decompress (invalid format) will be removed from the message.
 If the message results in zero parts it is skipped entirely.`,