+++ lib/processor/json.go
@@ -44,12 +44,6 @@ overwrites the previous contents with the new value.
 
 If the path is empty or "." the root of the data will be targeted.
 
-If the list of target parts is empty the processor will be applied to all
-message parts. Part indexes can be negative, and if so the part will be selected
-from the end counting backwards starting from -1. E.g. if part = -1 then the
-selected part will be the last part of the message, if part = -2 then the part
-before the last element with be selected, and so on.
-
 This processor will interpolate functions within the 'value' field, you can find
 a list of functions [here](../config_interpolation.md#functions).