+++ lib/processor/merge_json.go
@@ -36,13 +36,7 @@ func init() {
 Parses selected message parts as JSON blobs, attempts to merge them into one
 single JSON value and then writes it to a new message part at the end of the
 message. Merged parts are removed unless ` + "`retain_parts`" + ` is set to
-true.
-
-If the list of target parts is empty the processor will be applied to all
-message parts. Part indexes can be negative, and if so the part will be selected
-from the end counting backwards starting from -1. E.g. if part = -1 then the
-selected part will be the last part of the message, if part = -2 then the part
-before the last element with be selected, and so on.`,
+true.`,
 	}
 }